requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.2
python-dotenv==1.0.0
Werkzeug==3.1.3
